# MicroServicios/Products/src/main/resources/static/swagger.yaml
openapi: 3.0.1
info:
  title: "Products API"
  version: "1.0.0"
paths:
  /api/products:
    get:
      summary: Get Products
      description: Returns a list of products
      responses:
        '200':
          description: Product List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create product
      description: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Mistake
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/products/{id}:
    put:
      summary: Update product
      description: Update an existing product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      summary: Delete product
      description: Delete an existing product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successfully removed
  /api/products/find/{id}:
    get:
      summary: Check product
      description: Check if a product exists by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Dell XPS 8950 - Gaming Desktop"
        price:
          type: number
          format: double
          example: 1300
        quantity:
          type: integer
          example: 2